move form to insert entry page
