+++ app/routes/entries.tsx
@@ -12,14 +12,16 @@ export async function loader({ request }: LoaderArgs) {
   return json({ entries })
 }
 
-export async function action({ request }: ActionArgs) {
-  const data = Object.fromEntries(await request.formData());
-  putEntries(data);
-  return redirect(`/entries/${data.inputBox}`);
-}
-
 export default function EntriesPage() {
   const data = useLoaderData<typeof loader>();
+
+  /*
+  {data.entries.map((e) => {
+        return (<p className="relative mt-20 bg-white p-3 md:mx-auto md:mt-36 md:flex md:flex-row md:justify-center">
+          {e.headword}
+        </p>);
+      })}
+  */
 
   return (
     <div className="relative">
@@ -28,26 +30,6 @@ export default function EntriesPage() {
       <Main>
         <Outlet />
       </Main>
-      <Form id="newForm" className=" bg-gray-400 relative mt-20 p-3 md:mx-auto md:mt-36 md:flex md:flex-col md:justify-center" method="post">
-        <label className="p-3">
-          insert data
-          <input className="m-3" type="text" name="inputBox" />
-        </label>
-        <label className="p-3">
-          insert more data
-          <input className="m-3" type="text" name="inputBox2" />
-        </label>
-        <button className="bg-white" type="submit" name="submitButton">
-          Submit
-        </button>
-      </Form>
-
-      {data.entries.map((e) => {
-        return (<p className="relative mt-20 bg-white p-3 md:mx-auto md:mt-36 md:flex md:flex-row md:justify-center">
-          {e.headword}
-        </p>);
-      })}
-
     </div>
   )
 }
